abstract deprecation logic to util
