+++ index.js
@@ -2,7 +2,7 @@
 
 const { resolve } = require('path');
 const args = require('yargs-parser')(process.argv.slice(2));
-const { createProject } = require('./src/utils');
+const createProject = require('./src/utils/create-project');
 
 const project = createProject();
 
@@ -16,10 +16,10 @@ const preset = args.preset || args.p;
 const migration = args.migration || args.m;
 
 if (preset === 'v35') {
-  require('./src/button-import-deprecated')(project);
+  require('./src/use-deprecated-button')(project);
   require('./src/use-deprecated-actionlist')(project);
 } else {
-  if (migration === 'deprecate-buttons') require('./src/button-import-deprecated')(project);
+  if (migration === 'deprecate-buttons') require('./src/use-deprecated-button')(project);
   else if (migration === 'use-deprecated-actionlist') require('./src/use-deprecated-actionlist')(project);
   else console.log('Migration not found! Did you mean `npx primer-react-migrate src -m deprecate-buttons`');
 }
