+++ src/use-deprecated-button.js
@@ -1,9 +1,26 @@
 const { SyntaxKind } = require('ts-morph');
 const updateImportDeclaration = require('./utils/change-import-to-deprecated');
 
-const actionListImportNames = ['ActionList', 'ActionListProps', 'GroupProps', 'ItemProps'];
+const buttonImportNames = [
+  'Button',
+  'ButtonProps',
+  'ButtonDanger',
+  'ButtonDangerProps',
+  'ButtonGroup',
+  'ButtonGroupProps',
+  'ButtonOutline',
+  'ButtonOutlineProps',
+  'ButtonPrimary',
+  'ButtonPrimaryProps',
+  'ButtonInvisible',
+  'ButtonInvisibleProps',
+  'ButtonTableList',
+  'ButtonTableListProps',
+  'ButtonClose',
+  'ButtonCloseProps'
+];
 
-const ignoreNewModuleSpecifier = 'ActionList2';
+const ignoreNewModuleSpecifier = 'Button2';
 
 const transform = (project) => {
   const sourceFiles = project.getSourceFiles();
@@ -11,7 +28,7 @@ const transform = (project) => {
   sourceFiles.forEach((sourceFile) => {
     try {
       sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(declaration, sourceFile, actionListImportNames, ignoreNewModuleSpecifier);
+        declaration = updateImportDeclaration(declaration, sourceFile, buttonImportNames, ignoreNewModuleSpecifier);
       });
 
       // save source back to file
