+++ src/utils/create-project.js
@@ -8,7 +8,4 @@ const createProject = () => {
   });
 };
 
-// remove quotes from string so that they can be used as StringLiteral
-const convertToStringLiteral = (string) => string.replaceAll(`'`, ``).replaceAll(`"`, ``);
-
-module.exports = { createProject, convertToStringLiteral };
+module.exports = createProject;
