+++ src/utils/string-literal.js
@@ -1,0 +1,4 @@
+// remove quotes from string so that they can be used as StringLiteral
+const convertToStringLiteral = (string) => string.replaceAll(`'`, ``).replaceAll(`"`, ``);
+
+module.exports = { convertToStringLiteral };
